Fix most of the warnings
