+++ Habit-Calendar/Habit-Calendar-Tests/.swiftlint.yml
@@ -1,0 +1,3 @@
+disabled_rules:
+ - type_body_length
+ - file_length
